+++ client/src/Pages/Forum.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Navbar from '../components/Navbar'; 
+
+function Forum() {
+    return (
+        <div className="Forum">
+            <Navbar />
+           <h1>Forum</h1>
+        </div>
+    );
+}
+
+
+export default Forum;
