+++ client/src/Pages/Quiz.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../App.css';
+import Navbar from '../components/Navbar'; 
+
+function Quiz() {
+    return (
+        <div className="Quiz">
+            <Navbar />
+           
+        </div>
+    );
+}
+
+
+export default Quiz;
