+++ client/src/Pages/Todo.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Navbar from '../components/Navbar'; 
+
+function Todo() {
+    return (
+        <div className="Todo">
+            <Navbar />
+           
+        </div>
+    );
+}
+
+
+export default Todo;
