+++ client/src/Pages/Video.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Navbar from '../components/Navbar'; 
+
+function Video() {
+    return (
+        <div className="Video">
+            <Navbar />
+            
+        </div>
+    );
+}
+
+
+export default Video;
