+++ client/src/index.js
@@ -2,6 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import Home from './App';
+import Forum from './Pages/Forum';
+import Video from './Pages/Video';
+import Quiz from './Pages/Quiz';
+// import Todo from './Pages/Todo';
+
 import reportWebVitals from './reportWebVitals';
 import {BrowserRouter, Routes, Route} from 'react-router-dom'
 
@@ -10,9 +15,9 @@ root.render(
   <BrowserRouter>
     <Routes>
       <Route path="/" element={<Home/>}/>
-      <Route path="/video" element={<Home/>}/> {/*Add the required element to be rendered in the element attribute (according to the path specified)*/}
-      <Route path="/quiz" element={<Home/>}/>
-      <Route path="/forum" element={<Home/>}/>
+      <Route path="/video" element={<Video/>}/> {/*Add the required element to be rendered in the element attribute (according to the path specified)*/}
+      <Route path="/quiz" element={<Quiz/>}/>
+      <Route path="/forum" element={<Forum/>}/>
       <Route path="/todo" element={<Home/>}/>
       <Route path="/profile" element={<Home/>}/>
     </Routes>
